refactor: structure update
