+++ src/components/index.tsx
@@ -1,4 +1,2 @@
 export * from "./ConstraintSelect";
 export * from "./VersionSelect";
-export * from "./Channels";
-export * from "./RequestedPackages";
