+++ src/features/channels/components/Channel.tsx
